+++ GeometryBasicsCore/Project.toml
@@ -4,10 +4,12 @@ authors = ["SimonDanisch <sdanisch@gmail.com>"]
 version = "1.0.0"
 
 [deps]
+IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
+IterTools = "1.3.0"
 StaticArrays = "0.12, 1.0"
 julia = "1.6"
