+++ Project.toml
@@ -7,7 +7,6 @@ version = "0.4.3"
 EarCut_jll = "5ae413db-bbd1-5e63-b57d-d24a61df00f5"
 GeoInterface = "cf35fbd7-0cd7-5166-be24-54bfbe79505f"
 GeometryBasicsCore = "609001ca-df33-43ad-af5e-98eca4304ed0"
-IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
@@ -18,7 +17,6 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 EarCut_jll = "2"
 GeoInterface = "1.0.1"
 GeometryBasicsCore = "1"
-IterTools = "1.3.0"
 Reexport = "1"
 StaticArrays = "0.12, 1.0"
 StructArrays = "0.6"
