+++ src/metadata.jl
@@ -31,37 +31,6 @@ end
 
 getcolumn(t, colname::Symbol) = getcolumns(t, colname)[1]
 
-"""
-    MetaType(::Type{T})
-
-Returns the Meta Type corresponding to `T`
-E.g:
-```julia
-MetaType(Point) == PointMeta
-```
-"""
-MetaType(::Type{T}) where {T} = error("No Meta Type for $T")
-
-"""
-    MetaFree(::Type{T})
-
-Returns the original type containing no metadata for `T`
-E.g:
-```julia
-MetaFree(PointMeta) == Point
-```
-"""
-MetaFree(::Type{T}) where {T} = error("No meta free Type for $T")
-
-"""
-    meta(x::MetaObject)
-
-Returns the metadata of `x`
-"""
-meta(x::T) where {T} = error("$T has no meta!")
-
-metafree(x::T) where {T} = x
-
 macro meta_type(name, mainfield, supertype, params...)
     MetaName = Symbol("$(name)Meta")
     field = QuoteNode(mainfield)
@@ -102,12 +71,12 @@ macro meta_type(name, mainfield, supertype, params...)
             return $MetaName{$(params_sym...),ST,Names,Types}
         end
 
-        GeometryBasics.MetaFree(::Type{<:$MetaName{$(params_sym...),Typ}}) where {$(params_sym...), Typ<:$supertype{$(params_sym...)} } = Typ
-        GeometryBasics.MetaFree(::Type{<:$MetaName}) = $name
-        GeometryBasics.metafree(x::$MetaName) = getfield(x, :main)
-        GeometryBasics.metafree(x::AbstractVector{<:$MetaName}) = getproperty(x, $field)
-        GeometryBasics.meta(x::$MetaName) = getfield(x, :meta)
-        GeometryBasics.meta(x::AbstractVector{<:$MetaName}) = getproperty(x, :meta)
+        GeometryBasicsCore.MetaFree(::Type{<:$MetaName{$(params_sym...),Typ}}) where {$(params_sym...), Typ<:$supertype{$(params_sym...)} } = Typ
+        GeometryBasicsCore.MetaFree(::Type{<:$MetaName}) = $name
+        GeometryBasicsCore.metafree(x::$MetaName) = getfield(x, :main)
+        GeometryBasicsCore.metafree(x::AbstractVector{<:$MetaName}) = getproperty(x, $field)
+        GeometryBasicsCore.meta(x::$MetaName) = getfield(x, :meta)
+        GeometryBasicsCore.meta(x::AbstractVector{<:$MetaName}) = getproperty(x, :meta)
 
         function GeometryBasics.meta(main::$supertype{$(params_sym...)};
                                      meta...) where {$(params...)}
@@ -209,72 +178,6 @@ Base.size(x::MultiPolygonMeta) = size(metafree(x))
 Base.getindex(x::MeshMeta, idx::Int) = getindex(metafree(x), idx)
 Base.size(x::MeshMeta) = size(metafree(x))
 
-"""
-
-    MetaT(geometry, meta::NamedTuple)
-    MetaT(geometry; meta...)
-    
-Returns a `MetaT` that holds a geometry and its metadata
-
-`MetaT` acts the same as `Meta` method.
-The difference lies in the fact that it is designed to handle
-geometries and metadata of different/heterogeneous types.
-
-eg: While a Point MetaGeometry is a `PointMeta`, the MetaT representation is `MetaT{Point}`
-The downside being it's not subtyped to `AbstractPoint` like a `PointMeta` is.
-
-Example:
-```julia
-julia> MetaT(Point(1, 2), city = "Mumbai")
-MetaT{Point{2,Int64},(:city,),Tuple{String}}([1, 2], (city = "Mumbai",))
-```
-"""
-struct MetaT{T,Names,Types}
-    main::T
-    meta::NamedTuple{Names,Types}
-end
-
-MetaT(x; kwargs...) = MetaT(x, values(kwargs))
-
-"""
-
-    metafree(x::MetaT)
-    metafree(x::Array{MetaT})
-
-Free the MetaT from metadata
-i.e. returns the geometry/array of geometries
-"""
-function metafree(x::MetaT)
-    return getfield(x, :main)
-end
-metafree(x::AbstractVector{<:MetaT}) = map(metafree, x)
-
-"""
-
-    meta(x::MetaT)
-    meta(x::Array{MetaT})
-    
-Returns the metadata of a `MetaT`
-"""
-function meta(x::MetaT)
-    return getfield(x, :meta)
-end
-meta(x::AbstractVector{<:MetaT}) = map(meta, x)
-
-# helper methods
-function Base.getproperty(x::MetaT, field::Symbol)
-    return if field == :main
-        metafree(x)
-    elseif field == :meta
-        meta(x)
-    else
-        getproperty(meta(x), field)
-    end
-end
-
-Base.propertynames(x::MetaT) = (:main, propertynames(meta(x))...)
-getnamestypes(::Type{MetaT{T,Names,Types}}) where {T,Names,Types} = (T, Names, Types)
-
 # explicitly give the "schema" of the object to StructArrays
 function StructArrays.staticschema(::Type{F}) where {F<:MetaT}
     T, names, types = getnamestypes(F)
@@ -300,6 +203,3 @@ function meta_table(main, meta::NamedTuple{names}) where {names}
     F = MetaT{eltype(main),names,eltypes}
     return StructArray{F}(; main=main, meta...)
 end
-
-Base.getindex(x::MetaT, idx::Int) = getindex(metafree(x), idx)
-Base.size(x::MetaT) = size(metafree(x))
